Get the build correctly placing the dependencies in the assembly.
Consolidated one otherwise duplicative resource.
