+++ server/monitor/src/main/java/org/apache/accumulo/monitor/Monitor.java
@@ -583,7 +583,7 @@ public class Monitor implements HighlyAvailableService {
     }
   }
 
-  protected static void fetchScans() throws Exception {
+  public static void fetchScans() throws Exception {
     if (instance == null)
       return;
     Connector c = context.getConnector();
